web/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pyaudio==0.2.14
requests==2.31.0
shazamio==0.8.1
mutagen==1.47.0
python-dotenv==1.0.0
yt-dlp==2023.12.30
pydub==0.25.1

